Share NDJSON range splitting
